move some events to where they belong
