+++ src/main/java/serverutils/handlers/ServerUtilitiesPlayerEventHandler.java
@@ -10,12 +10,9 @@ import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.ChatComponentText;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.IChatComponent;
-import net.minecraft.util.MathHelper;
-import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.World;
 import net.minecraftforge.event.entity.EntityEvent;
 import net.minecraftforge.event.entity.living.LivingAttackEvent;
@@ -46,7 +43,6 @@ import serverutils.lib.data.ForgeTeam;
 import serverutils.lib.data.Universe;
 import serverutils.lib.math.BlockDimPos;
 import serverutils.lib.math.ChunkDimPos;
-import serverutils.lib.math.MathUtils;
 import serverutils.lib.util.InvUtils;
 import serverutils.lib.util.ServerUtils;
 import serverutils.lib.util.StringUtils;
@@ -151,73 +147,6 @@ public class ServerUtilitiesPlayerEventHandler {
     public static void onEntityDamage(LivingAttackEvent event) {
         if (ServerUtilitiesConfig.world.disable_player_suffocation_damage && event.entity instanceof EntityPlayer
                 && (event.source == DamageSource.inWall)) {
-            event.setCanceled(true);
-        }
-    }
-
-    @SubscribeEvent(priority = EventPriority.HIGHEST)
-    public static void onEntityAttacked(AttackEntityEvent event) {
-        if (!ClaimedChunks.canAttackEntity(event.entityPlayer, event.target)) {
-            InvUtils.forceUpdate(event.entityPlayer);
-            event.setCanceled(true);
-        }
-    }
-
-    @SubscribeEvent(priority = EventPriority.HIGHEST)
-    public static void onPlayerInteraction(PlayerInteractEvent event) {
-        EntityPlayer player = event.entityPlayer;
-        boolean cancelled = false;
-
-        if (ServerUtilitiesConfig.world.isItemRightClickDisabled(player.getHeldItem())) {
-            cancelled = true;
-            if (!event.world.isRemote) {
-                player.addChatComponentMessage(
-                        new ChatComponentText(EnumChatFormatting.DARK_PURPLE + "Item is disabled!"));
-            }
-        }
-
-        int x = event.x;
-        int y = event.y;
-        int z = event.z;
-
-        if (event.action == PlayerInteractEvent.Action.RIGHT_CLICK_AIR) {
-            MovingObjectPosition lookPos = MathUtils.rayTrace(player, true);
-            if (lookPos != null) {
-                x = lookPos.blockX;
-                y = lookPos.blockY;
-                z = lookPos.blockZ;
-            } else {
-                x = MathHelper.floor_double(player.posX);
-                y = MathHelper.floor_double(player.posY);
-                z = MathHelper.floor_double(player.posZ);
-            }
-        }
-
-        if (!cancelled) {
-            cancelled = switch (event.action) {
-                case RIGHT_CLICK_AIR -> ClaimedChunks.blockItemUse(player, x, y, z);
-                case RIGHT_CLICK_BLOCK -> ClaimedChunks.blockBlockInteractions(player, x, y, z, 0);
-                case LEFT_CLICK_BLOCK -> ClaimedChunks.blockBlockEditing(player, x, y, z, 0);
-            };
-        }
-
-        if (cancelled) {
-            event.setCanceled(true);
-            InvUtils.forceUpdate(player);
-        }
-    }
-
-    @SubscribeEvent(priority = EventPriority.HIGHEST)
-    public static void onBlockBreak(BlockEvent.BreakEvent event) {
-        if (ClaimedChunks.blockBlockEditing(event.getPlayer(), event.x, event.y, event.z, 0)) {
-            event.setCanceled(true);
-        }
-    }
-
-    @SubscribeEvent(priority = EventPriority.HIGHEST)
-    public static void onBlockPlace(BlockEvent.PlaceEvent event) {
-        if (ClaimedChunks.blockBlockEditing(event.player, event.x, event.y, event.z, 0)) {
-            InvUtils.forceUpdate(event.player);
             event.setCanceled(true);
         }
     }
